WebApp u
Used for the demo in the ZIP File attached. There is a README with the settings and things you'll need to change.
JQuery, Dataiku and D3.js should be checked.
You will need to change the Webapp API key to the one in your webapp settings.
You will need to give that key appropriate permissions in the project security settings.
You will need to set the name of the scenario that you wish to run.
You will need to adapt the security settings to your needs and environment.

Note: This webapp code is shared as a starter code.
The code has been kept simple for demo purposes and different ways of doing the same things are included in different places by design.
This should not be considered ready for deployment and is for learning purposes only.
It will be up to the user to implement appropriate checks on any user input and adequate error handling.

Additional note: Writing in a database from a Rshiny web app is not possible today in an instance with Multi-User Security. However, our future version (5.1.5 or 5.2) will make it possible, the version is expected for July/August